+++ test/boarding_view_test.dart
@@ -9,19 +9,16 @@ main() {
   var canvas = querySelector('#canvas');
   var surface = new Surface(grid, canvas, withLines: false);
   surface.draw();
-  var circle1 = new Circle(surface, 200, 200, 40,
-    lineWidth: 4, inColor: 'yellow', outColor: 'brown');
+  var circle1 = new Circle(surface, x: 200, y:200, width: 80, lineWidth: 4, color: 'yellow', borderColor: 'brown');
   circle1.draw();
-  var rectangle1 = new Rect(surface, 400, 300, 80, 40);
+  var rectangle1 = new Rect(surface, x: 400, y: 300, width: 80, height: 40);
   rectangle1.draw();
-  var rectangle2 = new Rect(surface, 100, 300, 40, 60, inColor: 'green');
+  var rectangle2 = new Rect(surface, x: 100, y: 300, width: 40, height: 60, color: 'green');
   rectangle2.draw();
-  var square1 =
-    new Square(surface, 40, 500, 60, lineWidth: 8, inColor: 'orange');
+  var square1 = new Square(surface, x: 40, y: 500, width: 60, lineWidth: 8, color: 'orange');
   square1.draw();
-  var line1 = new Line(surface, 450, 380, 480, 560, width: 3, color: '#ff0000');
+  var line1 = new Line(surface, x: 450, y: 380, width: 480, height: 560, lineWidth: 3, color: 'blue');
   line1.draw();
-  var tag1 =
-      new Tag(surface, 'Start on DArt', 320, 520, size: 32, color: 'blue');
+  var tag1 = new Tag(surface, x: 320, y: 520, width: 32, text: 'Start on DArt', color: 'red');
   tag1.draw();
 }