+++ CHANGELOG.md
@@ -5,6 +5,7 @@ based on [Semantic Versioning](http://semver.org/)
 *0.1.3* 2015-02-19
 
 + add moving pieces in the model
++ add abstract class Shape in the view
 + add Chaos example
 
 *0.1.2* 2015-01-27